Move "Changelog diff is:" out of the first line of the commit message.
When viewing the created commit messages with something like "git log
--oneline", or when using the first line of a commit message as an email,
it looks a little odd to see something like this:

"Checking in changes prior to tagging of version 1.00.  Changelog diff is:"

By moving the "Changelog diff is:" out of the first line of the commit
message, the use cases above no longer look odd.